requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
inference-sdk>=0.9.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
reportlab>=4.0.0
plotly>=5.17.0